Refactor immunity_results/ to take multiple illnesses with same pattern. Change list of illnesses to a dictionary. Under the key for each illness is a dictionary with an 'immunity' key whose value is the function for that illness. The immunity function is passed all the dict in the session dict under the illness' key, as keyword arguments.
A probability of 'Unknown.' and the immunity_results_error_message are
passed to the template in case of Type/Value errors, so that an error
for one illness does not prevent the results page from rendering.

These results are packed into a results_dict with results under each
illness' name key. The results dict is unpacked as kwargs when passed to
render_template, allowing for varied keys to be passed to templates
for each illness if desired.