Add tests using configuration settings to test just Measles component.
This necessitated a importlib.reload of the ImmunityDataEntryForm object
as these tests follow the fuller tests, which set all the illnesses on
the form object, and they appear to otherwise persist across tests.

Renaming the test script such that this script testing measles was
tested first fixed the problem, but is obviously not a solution to be
relied on going forward.